refactor(AboutDialog): separate Credits components and styles
