+++ assets/scripts/dialogs/About/Credits.scss
@@ -1,0 +1,30 @@
+.credits-team {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  grid-auto-flow: row;
+  grid-gap: 1.5em;
+}
+
+.credits-team-past {
+  grid-template-columns: 1fr 1fr 1fr 1fr;
+}
+
+.credits-container {
+  display: flex;
+  margin-top: 1.5em;
+
+  > div {
+    flex-grow: 1;
+  }
+
+  h4 {
+    margin-top: 1em;
+    margin-bottom: 0.25em;
+    color: gray;
+    font-style: italic;
+  }
+
+  i {
+    color: gray;
+  }
+}
