+++ assets/scripts/dialogs/About/TeamMember.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import './TeamMember.scss'
 
 function TeamMember (props) {
   const { name, title, mugshotFile, url } = props
@@ -9,15 +10,15 @@ function TeamMember (props) {
     style.backgroundImage = `url('/images/team/${mugshotFile}')`
   }
 
-  const displayName = (url) ? (
-    <a target="_blank" rel="noopener noreferrer" href={url}>{name}</a>
-  ) : name
+  const displayName = (url)
+    ? <a target="_blank" rel="noopener noreferrer" href={url}>{name}</a>
+    : name
 
   return (
-    <div className="about-dialog-team-member">
-      <div className="about-dialog-team-mugshot" style={style} />
-      <span className="about-team-name">{displayName}</span>
-      <span className="about-team-title">{title}</span>
+    <div className="team-member">
+      <div className="team-member-mugshot" style={style} />
+      <span className="team-member-name">{displayName}</span>
+      <span className="team-member-title">{title}</span>
     </div>
   )
 }
