+++ assets/scripts/dialogs/About/TeamMember.scss
@@ -1,0 +1,46 @@
+.team-member {
+  text-align: center;
+}
+
+.team-member-name {
+  display: block;
+  margin-top: 0.5em;
+  font-size: 1.1em;
+  font-weight: bold;
+  line-height: 1.2;
+
+  .credits-team-past & {
+    font-size: 1em;
+  }
+
+  a {
+    color: black;
+    text-decoration: none;
+  }
+}
+
+.team-member-title {
+  display: block;
+  color: gray;
+  font-style: italic;
+  line-height: 1.2;
+  margin-top: 0.25em;
+}
+
+.team-member-mugshot {
+  max-width: 100%;
+  border-radius: 6px;
+
+  /* Placeholder image in case a mugshot image is not provided */
+  background-image: url('../../../images/avatar.png');
+  background-size: cover;
+  background-position: 50% 50%;
+  background-repeat: no-repeat;
+
+  /* Forces the mugshot div height to be a perfect square, no matter its width */
+  &::after {
+    content: '';
+    display: block;
+    padding-bottom: 100%;
+  }
+}
