+++ assets/scripts/dialogs/AboutDialog.scss
@@ -43,6 +43,11 @@
   }
 }
 
+.about-dialog-content {
+  display: flex;
+  margin-bottom: 2em;
+}
+
 .about-dialog-left {
   flex-basis: 240px;
   margin-right: 40px;
@@ -88,21 +93,6 @@
   }
 }
 
-.about-dialog-team {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  grid-auto-flow: row;
-  grid-gap: 1.5em;
-}
-
-.about-dialog-team-past {
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-}
-
-.about-dialog-team-member {
-  text-align: center;
-}
-
 .about-dialog-sponsors {
   margin-top: -1em !important;
 
@@ -120,73 +110,4 @@
     margin-left: 1.5em;
     margin-right: 0;
   }
-}
-
-.about-dialog-content {
-  display: flex;
-  margin-bottom: 2em;
-}
-
-.about-team-name {
-  display: block;
-  margin-top: 0.5em;
-  font-size: 1.1em;
-  font-weight: bold;
-  line-height: 1.2;
-
-  .about-dialog-team-past & {
-    font-size: 1em;
-  }
-
-  a {
-    color: black;
-    text-decoration: none;
-  }
-}
-
-.about-team-title {
-  display: block;
-  color: gray;
-  font-style: italic;
-}
-
-.about-dialog-credits-container {
-  display: flex;
-  margin-top: 1.5em;
-}
-
-.about-dialog-credits-container > div {
-  flex-grow: 1;
-}
-
-.about-dialog-credits-container h4 {
-  margin-top: 1em;
-  margin-bottom: 0.25em;
-  color: gray;
-  font-style: italic;
-}
-
-.about-dialog-credits-container i {
-  color: gray;
-}
-
-.about-dialog-team img {
-  max-width: 100%;
-  border-radius: 6px;
-}
-
-.about-dialog-team-mugshot {
-  max-width: 100%;
-  border-radius: 6px;
-  background-image: url('../../images/avatar.png');
-  background-size: cover;
-  background-position: 50% 50%;
-  background-repeat: no-repeat;
-}
-
-/* Forces the mugshot div height to be a perfect square, no matter its width */
-.about-dialog-team-mugshot::after {
-  content: '';
-  display: block;
-  padding-bottom: 100%;
 }
